# Use a larger default h2 connection window