The default connection window (65535) is fairly small and limited in the
context of many applications.

grpc-go, for instance, uses a larger default value of 1048560 (slightly
less than 1MB). We now use this same default.

Fixes #638