@olix0r 2018-03-28
we may not want to merge this with the patch in Cargo.toml