@briansmith 2018-03-28
> grpc-go, for instance, uses a larger default value of 1048560 (slightly
> less than 1MB). We now use this same default.

How did you find this out? I cannot it in the source code and in fact when I looked at the source code it looks like it is 65535 in grpc-go master as of today: https://github.com/grpc/grpc-go/blob/f72b28a6d170d0938afd483331ba4e5d733785ae/transport/control.go#L34-L36

I read some of the grpc-go discussion regarding optimizing the connection window size in and a lot of it was driven by the desire to send very large messages, O(~1MB), over a high-latency connection (from Google's Austrailia data centers to to other continents w/ > ~100ms latency. That doesn't seem to fit the current expectations of Conduit workloads, which are expected to be primarily on low-latency networks. That's not to say I'm against this change, but I think we need more justification than "grpc-go does it." 

Downsides of a very large window I see are:
* Hugely increased memory usage per connection. If this is the default configuration we won't be able to reasonably claim to consume less than 10MB of RSS. Basically we'd have to take the memory utilization numbers out of the marketing, but that's one of the best numbers we have.
* Everything that bufferbloat causes. In particular this seems like it will interfere badly with load balancing for short messages since messages will get queued. e.g. ~1,000 1kb messages could be stuck in the buffer. In reality if hundreds of messages have been buffered we probably need to switch to a different server. That is, this seems like it has bad interactions with load balancing. (grpc-go doesn't do load balancing by default so it doesn't matter for them.)

Also, should we only be doing this for gRPC and not all h2 connections?

In the end grpc-go implemented a dynamic mechanism which is exactly what I think we will eventually need to. The question is what's a better interim default, small or huge or somewhere in the middle?

What benchmark are we using to measure this?