@olix0r 2018-03-29
> How did you find this out? I cannot it in the source code and in fact when I looked at the source code it looks like it is 65535 in grpc-go master as of today

This is used to configure the _stream_ window size, but not the _connection_ window size.

And so, this doesn't quite cause the buffer bloat one might expect, since stream windows are still limited. So the effect of this change mostly benefits use cases with concurrent streams.

I knew to look at this from working on Linkerd...  I discovered grpc-go's behavior by looking at `tshark` output -- a connection window update is sent by grpc-go peers at initialization.

in reality, as you mention, grpc-go uses 'bdp' to discover connection window settings dynamically.

having done the work to upgrade h2 to use these new settings, i see an improvement, but only a minor one.

On my laptop, running tests with 10 concurrent streams on a single connection, sending 100000 requests, each with a ~50000B response payload:

master runs at about 1800rps / 90MBps
```
:; release/darwin/strest-grpc-client -totalRequests 100000 -streams 10 -lengthPercentiles '1=50000,100=50000' -address localhost:4143
2018-03-29T00:14:05Z 786.2MB  16646/0 10s L:   0 [ 10  25 ]  333 J:   0   0
2018-03-29T00:14:15Z 844.0MB  17891/0 10s L:   0 [  9  18 ]  134 J:   0   0
2018-03-29T00:14:25Z 932.1MB  19751/0 10s L:   0 [  8  14 ]  122 J:   0   0
2018-03-29T00:14:35Z 831.6MB  17609/0 10s L:   0 [  9  19 ]  182 J:   0   0
2018-03-29T00:14:45Z 888.2MB  18805/0 10s L:   0 [  8  17 ]  153 J:   0   0
2018-03-29T00:14:55Z 437.8MB   9298/0 10s L:   0 [  9  19 ]   58 J:   0   0
{
  "good": 100000,
  "bad": 0,
  "bytes": 4949150000,
  "latency": {
    "50": 4,
    "95": 9,
    "99": 18,
    "999": 333
  },
  "jitter": {
    "50": 0,
    "95": 0,
    "99": 0,
    "999": 0
  }
}
```

on this branch, we see modest improvements (notably: better tail latencies)
```
:; release/darwin/strest-grpc-client -totalRequests 100000 -streams 10 -lengthPercentiles '1=50000,100=50000' -address localhost:4143
2018-03-29T00:15:56Z 926.7MB  19668/0 10s L:   0 [  7  12 ]   55 J:   0   0
2018-03-29T00:16:06Z 844.0MB  17877/0 10s L:   0 [  9  16 ]   70 J:   0   0
2018-03-29T00:16:16Z 929.4MB  19676/0 10s L:   0 [  8  11 ]   26 J:   0   0
2018-03-29T00:16:26Z 740.2MB  15657/0 10s L:   0 [ 11  23 ]   55 J:   0   0
2018-03-29T00:16:36Z 894.6MB  18955/0 10s L:   0 [  8  12 ]   30 J:   0   0
2018-03-29T00:16:46Z 385.7MB   8167/0 10s L:   0 [  8  13 ]   23 J:   0   0
{
  "good": 100000,
  "bad": 0,
  "bytes": 4950050000,
  "latency": {
    "50": 4,
    "95": 8,
    "99": 15,
    "999": 70
  },
  "jitter": {
    "50": 0,
    "95": 0,
    "99": 0,
    "999": 0
  }
}
```