@briansmith 2018-03-29
> This is used to configure the stream window size, but not the connection window size.

There used to be an initial connection window size variable, but it was removed in https://github.com/grpc/grpc-go/commit/d69dedd8b711c7d9875e0b9bf6c5c606a5a2f0cc#diff-f89d2e7360aeaf72f07c56cdfd03b00cL34

So, I think this variable is both the initial stream window size and initial connection window size. See: https://github.com/grpc/grpc-go/commit/d69dedd8b711c7d9875e0b9bf6c5c606a5a2f0cc#diff-3081f1aaaf958d7d5e00ccec7a62ea6fR197

The setting for the initial connection window size used to be 1024 * 1024 but that was reverted Mar 25, 2015, IIUC:
https://github.com/grpc/grpc-go/commit/310d8c389b1626a05ce048f4b7c3106e12af7fb7

Let me know if I'm misunderstanding something.