@olix0r 2018-03-29
@briansmith interestingly, i still see this behavior in strest-grpc, which uses [grpc-go 1.7.3] (https://github.com/BuoyantIO/strest-grpc/blob/master/Gopkg.toml#L48-L50) (released Dec 5 1017).  So, we're missing part of the picture.