@olix0r 2018-03-29
Note that grpc-go transports support a [server](https://github.com/grpc/grpc-go/blob/57640c0e6fd5e562b3a53f6b72738759165c80e1/transport/transport.go#L474) and [client](https://github.com/grpc/grpc-go/blob/57640c0e6fd5e562b3a53f6b72738759165c80e1/transport/transport.go#L507) `InitialConnWindowSize` -- we don't configure this anywhere explicitly in strest...