@olix0r 2018-03-29
Ah hah! it turns out that i've been running an ancient version of strest.  Now that I've updated, I no longer see these window updates from grpc-go.  Craziness resolved.

I think we should discard this for now and revisit later if needed.